backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.13.1
Pillow==10.2.0
email-validator==2.1.0.post1
aiofiles==23.2.1
qrcode[pil]==7.4.2
greenlet

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
aiosqlite==0.19.0
